Workaround for new style netflix using ratewizard
Using the ratingswizard, only a URL with get-requests is needed to
import a rating.

Now, you need a file called 'account.json' , in which the accountid and
authentication url are put. These can be found on
http://www.netflix.com/RatingsWizard?skgr=false&wizst=2, by copying the
link address from one the rating stars (any).
The URL looks like this
http://www.netflix.com/SetRating?value=4&pval=4.0626845&widgetid=M70001989_XXXXXXXXXX&rtrnct=true&authURL=YYYYYYYYYYYYYYYYYYYYYYYYY&section=RECS

XXXX refers to the account id (you can skip anything from _
(underscore), if present.
YYYYYYY refers to the authentication url